# doc/swagger/justtalk.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "rpc_create_user.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "JustTalk"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/create_user": {
      "post": {
        "operationId": "JustTalk_CreateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCreateUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbCreateUserRequest"
            }
          }
        ],
        "tags": [
          "JustTalk"
        ]
      }
    },
    "/v1/delete_user": {
      "patch": {
        "operationId": "JustTalk_DeleteUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbDeleteUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbDeleteUserRequest"
            }
          }
        ],
        "tags": [
          "JustTalk"
        ]
      }
    },
    "/v1/get_user_info": {
      "patch": {
        "operationId": "JustTalk_GetUserInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbGetUserInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbGetUserInfoRequest"
            }
          }
        ],
        "tags": [
          "JustTalk"
        ]
      }
    },
    "/v1/login_user": {
      "post": {
        "operationId": "JustTalk_LoginUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbLoginUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbLoginUserRequest"
            }
          }
        ],
        "tags": [
          "JustTalk"
        ]
      }
    },
    "/v1/update_user": {
      "patch": {
        "operationId": "JustTalk_UpdateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbUpdateUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbUpdateUserRequest"
            }
          }
        ],
        "tags": [
          "JustTalk"
        ]
      }
    }
  },
  "definitions": {
    "pbCreateUserRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "avatarUrl": {
          "type": "string"
        },
        "friends": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "int64"
          }
        }
      }
    },
    "pbCreateUserResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "avatarUrl": {
          "type": "string"
        },
        "friendCount": {
          "type": "integer",
          "format": "int32"
        },
        "friends": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "int64"
          }
        },
        "createAt": {
          "type": "string",
          "format": "date-time"
        },
        "updateAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "pbDeleteUserRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "pbDeleteUserResponse": {
      "type": "object",
      "properties": {
        "response": {
          "type": "string"
        }
      }
    },
    "pbGetUserInfoRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "pbGetUserInfoResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/pbUserResponse"
        }
      }
    },
    "pbLoginUserRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "pbLoginUserResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/pbUserResponse"
        }
      }
    },
    "pbSayHelloResponse": {
      "type": "object",
      "properties": {
        "answer": {
          "type": "string"
        }
      }
    },
    "pbUpdateUserRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "avatarUrl": {
          "type": "string"
        },
        "friendCount": {
          "type": "integer",
          "format": "int32"
        },
        "friends": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "int64"
          }
        }
      }
    },
    "pbUpdateUserResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "avatarUrl": {
          "type": "string"
        },
        "friendCount": {
          "type": "integer",
          "format": "int32"
        },
        "friends": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "int64"
          }
        },
        "createAt": {
          "type": "string",
          "format": "date-time"
        },
        "updateAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "pbUserResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "avatarUrl": {
          "type": "string"
        },
        "friendCount": {
          "type": "integer",
          "format": "int32"
        },
        "friends": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "int64"
          }
        },
        "createAt": {
          "type": "string",
          "format": "date-time"
        },
        "updateAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}